transform shop_resource into a concern that can be extends in apps
